Implement GitHub webhook server and refactor API structure
