+++ api/web_hook/models/__init__.py
@@ -1,0 +1,1 @@
+# Models package for GitHub webhook data structures
