+++ api/web_hook/prompts/__init__.py
@@ -1,0 +1,1 @@
+# Prompts package for LLM prompt templates
