+++ api/web_hook/utils/__init__.py
@@ -1,0 +1,1 @@
+# Utilities package for helper functions and common operations
